Reports kirt

See merge request !128

+++ StagingManagementSystem/src/main/java/com/revature/controllers/rest/AssociateControllerImpl.java
@@ -1,5 +1,6 @@
 package com.revature.controllers.rest;
 
+import java.util.Collection;
 import java.util.Set;
 
 import javax.servlet.http.HttpSession;
@@ -17,7 +18,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.revature.entities.Associate;
+import com.revature.entities.Manager;
 import com.revature.services.AssociateService;
+import com.revature.services.TotalReport;
+import com.revature.services.TotalReport.TotalData;
 import com.revature.util.DataGeneration;
 
 @RestController
@@ -28,6 +32,9 @@ public class AssociateControllerImpl {
 	AssociateService associateService;
 	@Autowired
 	DataGeneration dataGen;
+	@Autowired
+	TotalReport totalReport;
+	
 
 	public AssociateControllerImpl(AssociateService associateService) {
 		super();
@@ -35,16 +42,24 @@ public class AssociateControllerImpl {
 	}
 
 	@PostMapping
-	public void addAssociate(@RequestBody Associate associate) {
+	public ResponseEntity addAssociate(@RequestBody Associate associate, HttpSession session) {
+		if(session.getAttribute("login_manager") == null){ // If you're not logged in as a manger..
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		}
 		associateService.add(associate);
+		return ResponseEntity.ok(null);
 	}
 	
 	 @PostMapping("/add/all")
-	  public void addAssociates(@RequestBody Set<Associate> associates) {
-	   for(Associate a : associates){
-       associateService.add(a);       
-	   }
-	  }
+	 public ResponseEntity addAssociates(@RequestBody Set<Associate> associates, HttpSession session) {
+		 if(session.getAttribute("login_manager") == null){ // If you're not logged in as a manger..
+			 return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		 }
+		 for(Associate a : associates){
+			 associateService.add(a);
+		 }
+		 return ResponseEntity.ok(null);
+	 }
 	 
 	 @GetMapping("/generate/mock-data")
 	 public void generateAssociateMockDate(){
@@ -52,30 +67,58 @@ public class AssociateControllerImpl {
 	 }
 
 	@DeleteMapping
-	public void deleteAssociate(@RequestBody Associate associate) {
-		associateService.delete(associate);
+	public ResponseEntity deleteAssociate(@RequestBody Associate associate, HttpSession session) {
+		if(session.getAttribute("login_manager") == null){ // If you're not logged in as a manger..
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		}
+	 	associateService.delete(associate);
+		return ResponseEntity.ok(null);
 	}
 
 	@PutMapping
 	public ResponseEntity<Object> updateAssociate(@RequestBody Associate associate, HttpSession session) {
 		Associate authenticatedAssociate = (Associate)session.getAttribute("login_associate");
+		Manager authenticatedManager = (Manager)session.getAttribute("login_manager");
 		
-		if (authenticatedAssociate == null) {
-			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		if (authenticatedAssociate != null) { // Associate edits their profile
+			// Now we block any changes we don't want, by cherry picking the associate information
+			// from the passed in associate into the session associate.
+			authenticatedAssociate.setSkills(associate.getSkills());
+			authenticatedAssociate.setPortfolioLink(associate.getPortfolioLink());
+			associateService.update(authenticatedAssociate);
+			return ResponseEntity.ok(null);
 		}
-		
-		associate.setCredential(authenticatedAssociate.getCredential());
-		associateService.update(associate);
-		return ResponseEntity.ok(null);
+		Manager manager = (Manager)session.getAttribute("login_manager");
+		if(manager != null){// We trust managers. A lot.
+			associate.setCredential(associateService.getById(associate.getId()).getCredential());
+			associateService.update(associate);
+			return ResponseEntity.ok(null);
+		}
+
+		return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
 	}
 
 	@GetMapping("/{id}")
-	public Associate getAssociate(@PathVariable long id) {
-		return associateService.getById(id);
+	public ResponseEntity<Associate> getAssociate(@PathVariable long id, HttpSession session) {
+	 	Associate associate = ((Associate)session.getAttribute("login_associate"));
+		if(session.getAttribute("login_manager") == null && (associate == null || associate.getId() != id)){ // If you're not logged in as a manger..
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		}
+	 	return ResponseEntity.ok(associateService.getById(id));
 	}
 
 	@GetMapping("/all")
-	public Set<Associate> findById() {
-		return associateService.getAll();
+	public ResponseEntity<Set<Associate>> getAllAssociates(HttpSession session) {
+		if(session.getAttribute("login_manager") == null){ // If you're not logged in as a manger..
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		}
+		return ResponseEntity.ok(associateService.getAll());
 	}
+	
+	@GetMapping(path="/totaldata")
+	public ResponseEntity<Collection<TotalData>> getAssocaites(){
+		return ResponseEntity.ok(totalReport.process(associateService.getAll()));
+	}
+	
 }
+
